-Skybox edges problem fixed (waiting to test on wii). -Particle scripts now supported, only missing planar affector. -Particle Manager created. -Various platform improvements. -Wii threads now supported with mutexes. -New loading screen class, now shows "loading..." instead of black screen. -Demos revised and improved.

+++ trunk/src/timer.cpp
@@ -31,7 +31,7 @@ void timer::reset()
 long timer::getMilliSeconds()
 {
 	#ifdef __WII__
-		s64 end = gettime();
+		long long end = gettime();
 
 		return ticks_to_millisecs(diff_ticks(start, end));
 	#else
